Migrated record-specific resources to be subroutes of /patientrecord
